open-lessons/pb-and-grpc.07.11.2024/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
protobuf = "^5.28.3"
grpcio = "^1.67.1"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.67.1"

